seems like working all buttons except showlist
